requirements.txt
fastapi>=0.68.0,<1.0.0
uvicorn>=0.15.0,<1.0.0
pydantic>=2.0.0,<3.0.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
tenacity==9.0.0
